requirements/base.txt
bpython==0.24
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.1
drf-spectacular==0.27.2
pytz==2024.1
boto3==1.34.122
djangorestframework-api-key==2.*
pybase62==1.0.0
clerk-sdk-python==0.1.0
PyJWT[crypto]==2.8.0
djangorestframework-simplejwt==5.3.0

# Google Drive API integration
google-api-python-client==2.183.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2

# Django AllAuth for OAuth
django-allauth==65.11.2

# Token encryption
cryptography>=41.0.0

# QR Code generation
qrcode[pil]==7.4.2
Pillow>=10.0.0

